Move more stuff into LockableObject for abstraction
