+++ src/main/resources/htm.mixins.json
@@ -5,12 +5,13 @@
   "compatibilityLevel": "JAVA_16",
   "mixins": [
     "HopperBlockEntityMixin",
-    "LockableContainerMixin",
     "PistonBlockMixin",
     "events.BlockItemMixin",
     "events.EnderDragonEntityMixin",
     "events.ExplosionImplMixin",
-    "events.WorldMixin"
+    "events.WorldMixin",
+    "lockable.ChiseledBookshelfBlockEntityMixin",
+    "lockable.LockableContainerMixin"
   ],
   "injectors": {
     "defaultRequire": 1
